// src/types/partials/HumanName.ts
 
 
// This file is auto-generated by generate_types so do not edit manually

import { TExtension } from '../partials/Extension';
import { TPeriod } from '../partials/Period';

export type THumanName = {
  id?: string;
  extension?: TExtension[];
  use?: string;
  text?: string;
  family?: string;
  given?: string[];
  prefix?: string[];
  suffix?: string[];
  period?: TPeriod;
};