// src/types/partials/Address.ts
 
 
// This file is auto-generated by generate_types so do not edit manually

import { TExtension } from '../partials/Extension';
import { TPeriod } from '../partials/Period';

export type TAddress = {
  id?: string;
  extension?: TExtension[];
  use?: string;
  type?: string;
  text?: string;
  line?: string[];
  city?: string;
  district?: string;
  state?: string;
  postalCode?: string;
  country?: string;
  period?: TPeriod;
};