// src/types/partials/PatientCommunication.ts
 
 
// This file is auto-generated by generate_types so do not edit manually

import { TExtension } from '../partials/Extension';
import { TCodeableConcept } from '../partials/CodeableConcept';

export type TPatientCommunication = {
  id?: string;
  extension?: TExtension[];
  modifierExtension?: TExtension[];
  language: TCodeableConcept;
  preferred?: boolean;
};